Revert mapbox-gl back to v0.29.0 until v0.32.0 is released
Currently mapbox-gl v0.30.0 and v0.31.0 won't display the map in Firefox if the canvas is larger than 2048px in width or height (common on screens with resolution of 2K or higher).

The problem was introduced by mapbox/mapbox-gl-js#3841, in mapbox/mapbox-gl-js#3915 the issue with Firefox was discovered and in mapbox/mapbox-gl-js#3986 the fix was released.

So until v0.32.0 is released (which will incorporate the fix), I suggest reverting back to v0.29.0, which works fine with Firefox on large screens.